requirements.txt
fastapi
uvicorn[standard]
python-dotenv
asyncpg
sqlalchemy
databases
aioredis
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
pydantic[email]
bcrypt==4.0.1